+++ tests/macauff/test_fit_astrometry.py
@@ -149,7 +149,7 @@ class TestAstroCorrection:
                     **_kwargs, ax_dimension=1, npy_or_csv='csv',
                     coord_or_chunk='coord', coord_system=e_or_g, pregenerate_cutouts=True)
         for pregen_cut in [2, 'x', 'true']:
-            with pytest.raises(ValueError, match="pregenerate_cutouts should either be 'True' or "):
+            with pytest.raises(ValueError, match="pregenerate_cutouts should either be 'None', 'True' or "):
                 AstrometricCorrections(
                     **_kwargs, ax_dimension=1, npy_or_csv='csv', pregenerate_cutouts=pregen_cut,
                     coord_or_chunk='coord', coord_system='equatorial')
@@ -179,11 +179,11 @@ class TestAstroCorrection:
         self.a_cat_name = 'store_data/a_cat{}{}.npy'
         self.b_cat_name = 'store_data/b_cat{}{}.npy'
         with pytest.raises(ValueError, match='a_cat_func must be given if pregenerate_cutouts '):
-            ac(self.a_cat_name, self.b_cat_name, a_cat_func=None, b_cat_func=None,
+            ac(a_cat_name=self.a_cat_name, b_cat_name=self.b_cat_name, a_cat_func=None, b_cat_func=None,
                tri_download=False, make_plots=True, make_summary_plot=True)
         with pytest.raises(ValueError, match='b_cat_func must be given if pregenerate_cutouts '):
-            ac(self.a_cat_name, self.b_cat_name, a_cat_func=self.fake_cata_cutout, b_cat_func=None,
-               tri_download=False, make_plots=True, make_summary_plot=True)
+            ac(a_cat_name=self.a_cat_name, b_cat_name=self.b_cat_name, a_cat_func=self.fake_cata_cutout,
+               b_cat_func=None, tri_download=False, make_plots=True, make_summary_plot=True)
         dd_params = np.load(os.path.join(os.path.dirname(__file__), 'data/dd_params.npy'))
         l_cut = np.load(os.path.join(os.path.dirname(__file__), 'data/l_cut.npy'))
         ax1_mids, ax2_mids = np.array([105], dtype=float), np.array([0], dtype=float)
@@ -203,6 +203,34 @@ class TestAstroCorrection:
             coord_or_chunk='coord', pos_and_err_indices=[[0, 1, 2], [0, 1, 2]], mag_indices=[3],
             mag_unc_indices=[4], mag_names=['W1'], best_mag_index=0, coord_system='equatorial',
             chunks=chunks, pregenerate_cutouts=True, n_r=2000, n_rho=2000, max_rho=40)
+        with pytest.raises(ValueError, match="a_cat and b_cat must either both be None or "):
+            ac(a_cat=None, b_cat=np.array([0]), a_cat_name=None, b_cat_name=None, a_cat_func=None,
+               b_cat_func=None, tri_download=False, make_plots=True, make_summary_plot=True)
+        with pytest.raises(ValueError, match="a_cat_name and b_cat_name must either both be None or "):
+            ac(a_cat=None, b_cat=None, a_cat_name=None, b_cat_name='text', a_cat_func=None,
+               b_cat_func=None, tri_download=False, make_plots=True, make_summary_plot=True)
+        with pytest.raises(ValueError, match="pregenerate_cutouts must be None if a_cat is not None"):
+            ac(a_cat=np.array([0]), b_cat=np.array([0]), a_cat_name=None, b_cat_name=None, a_cat_func=None,
+               b_cat_func=None, tri_download=False, make_plots=True, make_summary_plot=True)
+        ac.pregenerate_cutouts = None
+        with pytest.raises(ValueError, match="a_cat_func must be None if a_cat is not "):
+            ac(a_cat=np.array([0]), b_cat=np.array([0]), a_cat_name=None, b_cat_name=None,
+               a_cat_func=np.array([0]), b_cat_func=None, tri_download=False, make_plots=True,
+               make_summary_plot=True)
+        with pytest.raises(ValueError, match="b_cat_func must be None if b_cat is not "):
+            ac(a_cat=np.array([0]), b_cat=np.array([0]), a_cat_name=None, b_cat_name=None,
+               a_cat_func=None, b_cat_func=np.array([0]), tri_download=False, make_plots=True,
+               make_summary_plot=True)
+        with pytest.raises(ValueError, match="a_cat must not be None if pregenerate_cutouts is None."):
+            ac(a_cat=None, b_cat=None, a_cat_name=None, b_cat_name=None,
+               a_cat_func=None, b_cat_func=None, tri_download=False, make_plots=True,
+               make_summary_plot=True)
+        ac.pregenerate_cutouts = True
+        with pytest.raises(ValueError, match="a_cat and a_cat_name must not both be None or both not be "):
+            ac(a_cat=None, b_cat=None, a_cat_name=None, b_cat_name=None,
+               a_cat_func=None, b_cat_func=None, tri_download=False, make_plots=True,
+               make_summary_plot=True)
+
         self.npy_or_csv = 'npy'
         cat_args = (105.0, 0.0)
         if os.path.isfile(self.a_cat_name.format(*cat_args)):
@@ -211,24 +239,26 @@ class TestAstroCorrection:
             os.remove(self.b_cat_name.format(*cat_args))
         a_cat_func = None
         b_cat_func = None
+        ac.pregenerate_cutouts = True
         with pytest.raises(ValueError, match="If pregenerate_cutouts is 'True' all files must "
                            f"exist already, but {self.a_cat_name.format(*cat_args)} does not."):
-            ac(self.a_cat_name, self.b_cat_name, a_cat_func, b_cat_func,
-               tri_download=False, make_plots=True, make_summary_plot=True)
+            ac(a_cat_name=self.a_cat_name, b_cat_name=self.b_cat_name, a_cat_func=a_cat_func,
+               b_cat_func=b_cat_func, tri_download=False, make_plots=True, make_summary_plot=True)
         ax1_min, ax1_max, ax2_min, ax2_max = 100, 110, -3, 3
         self.fake_cata_cutout(ax1_min, ax1_max, ax2_min, ax2_max, *cat_args)
         with pytest.raises(ValueError, match="If pregenerate_cutouts is 'True' all files must "
                            f"exist already, but {self.b_cat_name.format(*cat_args)} does not."):
-            ac(self.a_cat_name, self.b_cat_name, a_cat_func, b_cat_func,
-               tri_download=False, make_plots=True, make_summary_plot=True)
+            ac(a_cat_name=self.a_cat_name, b_cat_name=self.b_cat_name, a_cat_func=a_cat_func,
+               b_cat_func=b_cat_func, tri_download=False, make_plots=True, make_summary_plot=True)
 
     @pytest.mark.remote_data
-    @pytest.mark.parametrize("npy_or_csv,coord_or_chunk,coord_system,pregenerate_cutouts,return_nm",
-                             [("csv", "chunk", "equatorial", True, False),
-                              ("npy", "coord", "galactic", False, True),
-                              ("npy", "chunk", "equatorial", False, False)])
+    @pytest.mark.parametrize("npy_or_csv,coord_or_chunk,coord_system,pregenerate_cutouts,return_nm,in_memory",
+                             [("csv", "chunk", "equatorial", True, False, False),
+                              ("npy", "coord", "galactic", None, True, True),
+                              ("npy", "chunk", "equatorial", False, False, False)])
     # pylint: disable-next=too-many-statements,too-many-branches
-    def test_fit_astrometry(self, npy_or_csv, coord_or_chunk, coord_system, pregenerate_cutouts, return_nm):
+    def test_fit_astrometry(self, npy_or_csv, coord_or_chunk, coord_system, pregenerate_cutouts, return_nm,
+                            in_memory):
         self.npy_or_csv = npy_or_csv
         dd_params = np.load(os.path.join(os.path.dirname(__file__), 'data/dd_params.npy'))
         l_cut = np.load(os.path.join(os.path.dirname(__file__), 'data/l_cut.npy'))
@@ -263,8 +293,8 @@ class TestAstroCorrection:
             pos_and_err_indices=[[0, 1, 2], [0, 1, 2]], mag_indices=[3], mag_unc_indices=[4],
             mag_names=['W1'], best_mag_index=0, coord_system=coord_system, chunks=chunks,
             pregenerate_cutouts=pregenerate_cutouts,
-            cutout_area=60 if not pregenerate_cutouts else None,
-            cutout_height=6 if not pregenerate_cutouts else None, n_r=2000, n_rho=2000, max_rho=40,
+            cutout_area=60 if pregenerate_cutouts is False else None,
+            cutout_height=6 if pregenerate_cutouts is False else None, n_r=2000, n_rho=2000, max_rho=40,
             return_nm=return_nm)
 
         if coord_or_chunk == 'coord':
@@ -296,13 +326,31 @@ class TestAstroCorrection:
                     np.array([[[-1, -1, -1, -1, -1], [-1, -2, -3, -4, -5]]], dtype=float))
             np.save('ac_save_folder/npy/m_sigs_array.npy', np.array([-1, 12], dtype=float))
             np.save('ac_save_folder/npy/n_sigs_array.npy', np.array([-1, 15], dtype=float))
+        if in_memory:
+            cat_args = (105.0, 0.0)
+            ax1_min, ax1_max, ax2_min, ax2_max = 100, 110, -3, 3
+            self.fake_cata_cutout(ax1_min, ax1_max, ax2_min, ax2_max, *cat_args)
+            self.fake_catb_cutout(ax1_min, ax1_max, ax2_min, ax2_max, *cat_args)
+            a_cat = np.load(self.a_cat_name.format(*cat_args))
+            b_cat = np.load(self.b_cat_name.format(*cat_args))
         if return_nm:
-            marray, narray, abc_array = ac(
-                self.a_cat_name, self.b_cat_name, a_cat_func, b_cat_func,
-                tri_download=False, make_plots=True, make_summary_plot=True)
+            if in_memory:
+                marray, narray, abc_array = ac(
+                    a_cat=a_cat, b_cat=b_cat, a_cat_name=None, b_cat_name=None, a_cat_func=None,
+                    b_cat_func=None, tri_download=False, make_plots=True, make_summary_plot=True)
+            else:
+                marray, narray, abc_array = ac(
+                    a_cat=None, b_cat=None, a_cat_name=self.a_cat_name, b_cat_name=self.b_cat_name,
+                    a_cat_func=a_cat_func, b_cat_func=b_cat_func, tri_download=False, make_plots=True,
+                    make_summary_plot=True)
         else:
-            ac(self.a_cat_name, self.b_cat_name, a_cat_func, b_cat_func,
-               tri_download=False, make_plots=True, make_summary_plot=True)
+            if in_memory:
+                ac(a_cat=a_cat, b_cat=b_cat, a_cat_name=None, b_cat_name=None, a_cat_func=None,
+                    b_cat_func=None, tri_download=False, make_plots=True, make_summary_plot=True)
+            else:
+                ac(a_cat=None, b_cat=None, a_cat_name=self.a_cat_name, b_cat_name=self.b_cat_name,
+                   a_cat_func=a_cat_func, b_cat_func=b_cat_func, tri_download=False, make_plots=True,
+                   make_summary_plot=True)
 
         if coord_or_chunk == 'coord':
             assert os.path.isfile('ac_save_folder/pdf/auf_fits_105.0_0.0.pdf')
@@ -392,12 +440,20 @@ class TestSNRMagRelation:
         with pytest.raises(ValueError, match="return_nm must either be True "):
             SNRMagnitudeRelationship(**_kwargs, ax_dimension=1, npy_or_csv='csv',
                                      coord_or_chunk='coord', coord_system='equatorial', return_nm='f')
+        smr = SNRMagnitudeRelationship(**_kwargs, ax_dimension=1, npy_or_csv='csv',
+                                       coord_or_chunk='coord', coord_system='equatorial', return_nm=False)
+        with pytest.raises(ValueError, match="One of b_cat and b_cat_name "):
+            smr(b_cat=None, b_cat_name=None)
+        with pytest.raises(ValueError, match="Only one of b_cat and b_cat_name "):
+            smr(b_cat=np.array([0]), b_cat_name='name')
 
-    @pytest.mark.parametrize("npy_or_csv,coord_or_chunk,coord_system,return_nm",
-                             [("csv", "chunk", "equatorial", True), ("npy", "coord", "galactic", False),
-                              ("npy", "chunk", "equatorial", False)])
+    @pytest.mark.parametrize("npy_or_csv,coord_or_chunk,coord_system,return_nm,in_memory",
+                             [("csv", "chunk", "equatorial", True, False),
+                              ("npy", "chunk", "galactic", True, True),
+                              ("npy", "coord", "galactic", False, False),
+                              ("npy", "chunk", "equatorial", False, False)])
     # pylint: disable-next=too-many-statements,too-many-branches
-    def test_snr_mag_relation_fit(self, npy_or_csv, coord_or_chunk, coord_system, return_nm):
+    def test_snr_mag_relation_fit(self, npy_or_csv, coord_or_chunk, coord_system, return_nm, in_memory):
         self.npy_or_csv = npy_or_csv
         # Flag telling us to test for the non-running of all sightlines,
         # but to leave pre-generated ones alone
@@ -459,10 +515,21 @@ class TestSNRMagRelation:
                     np.array([[[-1, -1, -1, -1, -1], [-1, -2, -3, -4, -5]]], dtype=float))
             np.save('ac_save_folder/npy/m_sigs_array.npy', np.array([-1, 12], dtype=float))
             np.save('ac_save_folder/npy/n_sigs_array.npy', np.array([-1, 15], dtype=float))
+        if in_memory:
+            cat_args = (chunks[0],)
+            ax1_min, ax1_max, ax2_min, ax2_max = 100, 110, -3, 3
+            t_a_c.fake_catb_cutout(ax1_min, ax1_max, ax2_min, ax2_max, *cat_args)
+            b_cat = [np.load(self.b_cat_name.format(*cat_args))]
         if return_nm:
-            abc_array, ax1_mids, ax2_mids = smr(self.b_cat_name, make_plots=True)
+            if in_memory:
+                abc_array = smr(b_cat=b_cat, make_plots=True)
+            else:
+                abc_array = smr(b_cat_name=self.b_cat_name, make_plots=True)
         else:
-            smr(self.b_cat_name, make_plots=True)
+            if in_memory:
+                smr(b_cat=b_cat, make_plots=True)
+            else:
+                smr(b_cat_name=self.b_cat_name, make_plots=True)
 
         if coord_or_chunk == 'coord':
             assert os.path.isfile('ac_save_folder/pdf/s_vs_snr_105.0_0.0.pdf')